Fixed some video mode issues and added simple text rendering.
The fullscreen opengl mode was drawing the widths of pixels
inconsistently. I fixed that. Now the 320x240 picture will
be scaled so that the height is closest to 240 and the
aspect ratio is equal to the aspect ratio of your original
video mode. The graphics will automatically be shifted right
and down (or up) to center the original view on widescreen
monitors.

The function draw_small_text() now works and will draw any
string containing characters less than 256.

I began rewriting the basic graphical backend so that instead
of having two drawing functions for each situation, there is
only two basic blitting functions (one for SDL and one for
opengl) and everything uses these depending on gl_flag. This
is certainly slower but unless I notice a problem the simplicity
gained is worth it.