Factored seq.c out of synth.c, rewrote both.
I copied the new algorithms out of another project I start a
few days ago. It seems to have a saner structure. I did get an
idea from the code that was replaced, looping can be simplified
by using control events for it instead of it being handled by
the sequencer without help. More on that later.