Began redesigning the stage
The stage module contains the state of the game world.
It has methods to load the world, draw the world, and do
collision tests with the (static) world.

Things that move around will be in the entity model, but
that is in another branch.

The basic shell of stage.c is in place, but is incomplete
and is untested. It uses the new list module, which is
incomplete. inner.c is the game module which updates, draws,
and handles events for the main game. It is expected to
utilize stage.c and entity.c.

The file game.c was unlinked and will be phased out, the
basic algorithms will reappear in another form throughout
inner.c stage.c and entity.c.