New mixer model in synth.c
I copied list.c and list.h from the stage branch to make it build.
A channel is a abstraction with pan, volume, a generator, and a
controller. A channel is expected to have many generators active
at once, for example a single polyphonic instrument can be a channel
and the entirety of sound effects is a channel.

Need to add a flag for a channel to mark it as sound effects or
music, so that we can have a master music and master sound volume.