Started rewriting the loader module.
game.c was replaced with iotest.c
This will be used to test the new loader.
We also might want to rethink graphics.c a little in this branch.

The main change is to loader.c, which no longer uses zziplib.
Don't like this library very much. In this branch it isn't even linked.
In the current state the program can't start because the console
letters can't load.

The goal is the a) write a zip file loader and b) to hookup this
to loader.c