Added list.c and list.h
These are actually copies of list.c and list.h from stage branch
commit 1fb39e

Was probably a mistake to add these in the beginning of that branch
since I was going to need them in the other ones.

I will now merge this into all topic branches, and master branch.