zip.c work.
Lots of changes to zip.c, mostly fixing problems and verifying
code. Error checking currently reports every file in the archive
to be not found. This is because when the archive is initially read
it runs a routine to insert the files found into a hash table. The
routine does nothing.

Next step is the write that routine.