Added audio output level probe. Amplified default instrument.
The probe can report in dB rms or dB peak. 0 is max, it looks like
-50dB rms is acceptable when I have the system volume on half.
We need to amplify instrument final output at max velocity so it
stays within this range. The default instrument needs to be changed
to use random phase since many notes may start on the same sample.