Stage loader checks for errors now. New file format defined.
But this commit stage.c loads the old format. An example of this
format is in data/zones/myzone.

During this commit I wrote a file with the new format in
data/zones/2zone. This is not represented in the repo but I am
uploading the new file now. Its untested and stage.c doesnt read
it yet.