Rewrite of the stage module.
This changes the format of the game somewhat. While the stage
basically is the same as it was superficially, the world no longer
consists of seamlessly linked zones. Instead you pass from stage
to stage via doors or other transistion mechanisms, and when you
beat the boss you go to a new zone completely.

I was having a hard time so I decided to simplify rather than
continue to agonize over what seemed like an overly complex and
unsuitable for implementation in C problem.

The inspiration for the new way is from lynx game gates of zendocon.
Rather than progress directly from level to level, there are secret
bosses and secret levels which may progress back to the main line.

The main format for the game play, and the dialog based plot has not
been changed.